requirements.txt
fastapi==0.85.1
uvicorn==0.18.2
tensorflow==2.11.0
pillow==9.2.0
numpy==1.23.4
requests==2.28.1
